backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
ccxt>=4.0.0
pandas>=2.2.0
pandas-ta>=0.3.14b
numpy>=1.26.0
redis>=5.0.0
apscheduler>=3.10.0
websockets>=12.0
httpx>=0.27.0
python-dotenv>=1.0.0
structlog>=24.0.0
orjson>=3.10.0
aiohttp>=3.9.0
